update code for homework 2
