+++ lib/logger/config.js
@@ -1,6 +1,5 @@
 import * as constants from './constants.js';
-// import {LOG_LEVEL, LOG_APPENDER} from './fileReader.js';
-import {logAppender, logLevel} from './fileReader.js';
+import {logAppender, logLevel} from './logConfig.js';
 
 
 const defaultConfig = {
